requirements.txt
fastapi
payments-py
python-dotenv
httpx